mico_store_BE/requirements.txt
djangorestframework
django-cors-headers
djangorestframework-simplejwt
asgiref
gunicorn

mongoengine
pymongo
dnspython
git+https://github.com/umutbozkurt/django-rest-framework-mongoengine.git
